T3_Recommandation/requirements.txt
psycopg2-binary
python-dotenv
pandas
numpy
scikit-learn
